made cash and total field long type

See merge request kount/third_party/kount-ris-java-sdk!19

+++ kount-ris-sdk/src/main/java/com/kount/ris/Inquiry.java
@@ -39,7 +39,7 @@ public class Inquiry extends Request {
 	 *            int Cash amount of any feasible goods
 	 * @return this
 	 */
-	public Inquiry setCash(int cash) {
+	public Inquiry setCash(long cash) {
 		this.params.put("CASH", String.valueOf(cash));
 		return this;
 	}
@@ -133,7 +133,7 @@ public class Inquiry extends Request {
 	 *            Transaction amount in lowest possible denomination of given currency
 	 * @return this
 	 */
-	public Inquiry setTotal(int total) {
+	public Inquiry setTotal(long total) {
 		this.params.put("TOTL", String.valueOf(total));
 		return this;
 	}
